# apps/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: قيّدها (Qeedha) API
  description: |
    منصة رصيد البقالة الممول بالتقسيط.
    النسخة 0.1.0 — المرحلة الأولى: عقود API الأساسية.
  version: 0.1.0
  contact:
    name: Qeedha Engineering
servers:
  - url: http://localhost:3001/v1
    description: Local development
  - url: https://api.qeedha.app/v1
    description: Production
security:
  - BearerAuth: []
paths:
...
  /transactions:
    get:
      tags: ["العمليات المالية"]
      summary: سجل عمليات العميل/البقالة
      parameters:
        - name: walletId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: قائمة العمليات
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
    post:
      tags: ["العمليات المالية"]
      summary: تنفيذ دفعة (QR أو OTP)
      description: |
        عملية دفع Idempotent باستخدام مفتاح Idempotency-Key. عند النجاح يتم خصم الرصيد وإنشاء
        قيد مزدوج في ledger_entries.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: تمت العملية
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "409":
          description: مفتاح Idempotency مكرر
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{transactionId}/refund:
    parameters:
      - $ref: "#/components/parameters/TransactionId"
    post:
      tags: ["العمليات المالية"]
      summary: استرجاع عملية
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "201":
          description: تم الاسترجاع
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /financing/applications/{applicationId}:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    get:
      tags: ["التمويل"]
      summary: تفاصيل طلب التمويل
      responses:
        "200":
          description: تفاصيل الطلب
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingApplication"
...
  /financing/webhooks/{provider}:
    post:
      tags: ["التمويل"]
      summary: Webhook وارد من مزود التمويل
      description: |
        يستقبل أحداث مزود التمويل (موافقة/رفض/سداد/تأخر/تعثر) بعد التحقق من التوقيع HMAC.
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [tabby, tamara, manual]
        - name: X-Idempotency-Key
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinancingWebhookPayload"
      responses:
        "200":
          description: تم استلام الحدث
...
  /admin/applications/{applicationId}/decide:
    parameters:
      - $ref: "#/components/parameters/ApplicationId"
    post:
      tags: ["لوحة الأدمن"]
      summary: قرار يدوي على طلب تمويل (ManualAdapter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDecisionRequest"
      responses:
        "200":
          description: تم اتخاذ القرار
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingApplication"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT صالح يُرسل في ترويسة Authorization.
  parameters:
...
    TransactionId:
      name: transactionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: مفتاح عالمي فريد لمنع تكرار العملية.
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        code:
          type: string
...
    FinancingPlan:
      type: string
      enum: [PAY_IN_1, PAY_IN_30, INSTALLMENTS_4]
...
    ApplicationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, CANCELLED, DEFAULTED]
...
    PaymentRequest:
      type: object
      required: [walletId, amount, method]
      properties:
        walletId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        method:
          $ref: "#/components/schemas/TransactionMethod"
        token:
          type: string
          description: توكن QR الموقّع (للـ QR) أو رمز OTP (للـ OTP)
        branchId:
          type: string
          format: uuid
        cashierId:
          type: string
          format: uuid
        metadata:
          type: object
    TransactionMethod:
      type: string
      enum: [QR, OTP, MANUAL]
    RefundRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
        amount:
          type: number
          format: decimal
          description: إذا لم يُرسل، يُسترجع المبلغ بالكامل.
    Transaction:
      type: object
      required: [id, walletId, type, amount, method, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          type: number
          format: decimal
        method:
          $ref: "#/components/schemas/TransactionMethod"
        status:
          $ref: "#/components/schemas/TransactionStatus"
        createdAt:
          type: string
          format: date-time
    TransactionType:
      type: string
      enum: [ACTIVATION, PURCHASE, REFUND, REVERSAL]
    TransactionStatus:
      type: string
      enum: [PENDING, COMPLETED, FAILED, REVERSED]
    TransactionListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    FinancingApplication:
      type: object
      required: [id, customerId, merchantId, providerId, amount, planType, status]
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        merchantId:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        planType:
          $ref: "#/components/schemas/FinancingPlan"
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        providerRef:
          type: string
        decidedAt:
          type: string
          format: date-time
...
    FinancingWebhookPayload:
      type: object
      required: [event, providerRef, occurredAt]
      properties:
        event:
          type: string
          enum:
            - application.approved
            - application.rejected
            - installment.paid
            - installment.overdue
            - application.defaulted
        providerRef:
          type: string
        applicationRef:
          type: string
        amount:
          type: number
          format: decimal
        occurredAt:
          type: string
          format: date-time
        signature:
          type: string
...
    AdminDecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
...
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer